+++ sql/table_asset.sql
@@ -2,7 +2,7 @@ drop table if exists daas.asset;
 
 create table if not exists daas.asset (
     id serial primary key,
-    fac_id bigint not null,
+    fac_id int not null,
     asset_id varchar(250) not null,
     sys_id varchar(250) null,
     create_ts timestamp,
